Utilize expand and a template sensor for security
I realized I can't just use groups because, for example, if we accidentally leave a door open it won't trip the alarm.

I then switched to a YAML list of entity IDs that should trip an alarm.

I discovered the expand function within templates, so decided I could make a template sensor that keeps track of tripped devices and counts them, and then update the automation to trip that alarm when that count increases.

The sensor also allows me to get the most recently tripped device, as well as *all* tripped devices, allowing me to better monitor the security of my home.